refactor: 개발자 토큰 서비스도 기존 컨벤션에 통합
