+++ src/main/java/com/techfork/auth/application/command/input/GenerateDeveloperTokenCommand.java
@@ -1,0 +1,6 @@
+package com.techfork.auth.application.command.input;
+
+public record GenerateDeveloperTokenCommand(
+        Long userId
+) {
+}
