+++ src/main/java/com/techfork/auth/application/command/result/DeveloperTokenResult.java
@@ -1,0 +1,9 @@
+package com.techfork.auth.application.command.result;
+
+import lombok.Builder;
+
+@Builder
+public record DeveloperTokenResult(
+        String developerToken
+) {
+}
